Move proto up a directory, add Makefile, fix Dockerfile.
