+++ cmd/eqmemsvc/svc.go
@@ -12,7 +12,7 @@ import (
 	"github.com/shiblon/entroq/qsvc"
 	"google.golang.org/grpc"
 
-	pb "github.com/shiblon/entroq/qsvc/proto"
+	pb "github.com/shiblon/entroq/proto"
 )
 
 var (
